pyproject.toml
[project]
name = "my-auto"
version = "0.1.0"
description = "Personal daily routine hub — multi-module orchestrator built on Claude Code Skills"
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "PyMuPDF>=1.24.0",
    "pdf2image>=1.17.0",
    "playwright>=1.45",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "responses>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests requiring real Obsidian CLI (deselect with '-m not integration')",
]
# Avoid module name collisions across modules (e.g. both auto-reading and
# auto-learning define test_models.py); each test file gets its own namespace.
addopts = "--import-mode=importlib"
